+++ Bencodex.Tests/Types/BooleanTest.cs
@@ -1,9 +1,9 @@
-﻿using Bencodex.Types;
+using Bencodex.Types;
 using Xunit;
+using static Bencodex.Tests.TestUtils;
 
 namespace Bencodex.Tests.Types
 {
-    // FIXME: Still some tests remain ValueTests.Boolean; they should come here.
     public class BooleanTest
     {
         private readonly Boolean _t = new Boolean(true);
@@ -24,10 +24,54 @@ namespace Bencodex.Tests.Types
         }
 
         [Fact]
+        public void Type()
+        {
+            Assert.Equal(ValueType.Boolean, _t.Type);
+            Assert.Equal(ValueType.Boolean, _f.Type);
+        }
+
+        [Fact]
+        public void EncodingLength()
+        {
+            Assert.Equal(1L, _t.EncodingLength);
+            Assert.Equal(1L, _f.EncodingLength);
+        }
+
+        [Fact]
         public void Fingerprint()
         {
             Assert.Equal(new Fingerprint(ValueType.Boolean, 1L, new byte[] { 1 }), _t.Fingerprint);
             Assert.Equal(new Fingerprint(ValueType.Boolean, 1L, new byte[] { 0 }), _f.Fingerprint);
+        }
+
+        [Fact]
+        public void Encode()
+        {
+            Codec codec = new Codec();
+            AssertEqual(
+                new byte[] { 0x74 },  // "t"
+                codec.Encode(_t)
+            );
+            AssertEqual(
+                new byte[] { 0x66 },  // "f"
+                codec.Encode(_f)
+            );
+        }
+
+        [Theory]
+        [InlineData(new object[] { false })]
+        [InlineData(new object[] { true })]
+        public void Inspect(bool loadAll)
+        {
+            Assert.Equal("true", _t.Inspect(loadAll));
+            Assert.Equal("false", _f.Inspect(loadAll));
+        }
+
+        [Fact]
+        public void String()
+        {
+            Assert.Equal("Bencodex.Types.Boolean true", _t.ToString());
+            Assert.Equal("Bencodex.Types.Boolean false", _f.ToString());
         }
     }
 }