+++ Bencodex.Tests/Types/NullTest.cs
@@ -1,15 +1,16 @@
 using Bencodex.Types;
 using Xunit;
+using static Bencodex.Tests.TestUtils;
 
 namespace Bencodex.Tests.Types
 {
-    // FIXME: Still some tests remain ValueTests.Null; they should come here.
     public class NullTest
     {
         [Fact]
         public void Value()
         {
             Assert.IsType<Null>(Null.Value);
+            Assert.Equal(default(Null), Null.Value);
         }
 
         [Fact]
@@ -34,6 +35,30 @@ namespace Bencodex.Tests.Types
         public void Type()
         {
             Assert.Equal(ValueType.Null, Null.Value.Type);
+        }
+
+        [Theory]
+        [InlineData(new object[] { false })]
+        [InlineData(new object[] { true })]
+        public void Inspect(bool loadAll)
+        {
+            Assert.Equal("null", Null.Value.Inspect(loadAll));
+        }
+
+        [Fact]
+        public void String()
+        {
+            Assert.Equal("Bencodex.Types.Null", Null.Value.ToString());
+        }
+
+        [Fact]
+        public void Encode()
+        {
+            Codec codec = new Codec();
+            AssertEqual(
+                new byte[] { 0x6e },  // "n"
+                codec.Encode(default(Null))
+            );
         }
     }
 }